[2002-03-28] 1) Prepare to separate 4 kinds of Descriptor handles. 2) Detect the transaction status more naturally. 3) Improve Parse Statement functionality for the use    of updatable cursors. 4) Improve updatable cursors. 5) Implement SQLGetDescField() and improve SQLColAttribute(). 6) etc.

+++ src/interfaces/odbc/dlg_specific.h
@@ -169,11 +169,11 @@ void		SetDlgStuff(HWND hdlg, const ConnInfo *ci);
 void		GetDlgStuff(HWND hdlg, ConnInfo *ci);
 
 int CALLBACK driver_optionsProc(HWND hdlg,
-				   WORD wMsg,
+				   UINT wMsg,
 				   WPARAM wParam,
 				   LPARAM lParam);
 int CALLBACK ds_optionsProc(HWND hdlg,
-			   WORD wMsg,
+			   UINT wMsg,
 			   WPARAM wParam,
 			   LPARAM lParam);
 #endif   /* WIN32 */
